Function to unlink entities on delimiter deletion
document.execCommand('unlink') doesn't look like it would perform very
similarly across browsers, so I'm being more explicit. If the entity
text boundaries no longer match the defined delimiters, the contents of
the anchor (in this case, the <span> saving cursor position and the text
node) will replace the anchor in the DOM tree.

This function (`removePrunedEntities`-- ripe for a rename), runs through
a DOM MutationObserver that observes every time an anchor's text is
edited.