Refine saveCursorPosition fn
saveCursorPosition works by adding a span where the cursor should go
once it all goes down. Previously this was always done after the current
selection-- however, this screwed things up when the span interrupted
the creation of an entity, which would happen if someone typed the
beginning delimiter to create it rather than the ending one.

    e.g. [|Alexander Berkman]    (| is the cursor)
         ^
         If this was typed last, the span would be placed afterwards and
         the regex would not match [<span></span>Alexander Berkman].

In the new function, if the character (in the future, characters) to the
left of the current selection exactly match the opening entity
delimiter, the span will be placed before that character/those
characters, and the cursor will be restored *before* the entity (meaning
it will have changed position-- that could be changed in
restoreCursorPosition in the future if need be.