Make restoreCursorPosition fn more accurate
In commit 629760f I updated the saveCurrentPosition function to insert a
placeholder span one character before the cursor if the immediately
previous character was a delimiter. This commit makes the
restoreCursorPosition function aware of that change. So, if that special
behavior ended up happening, the cursor would be restored one character
to the right after the span.

n.b. rangy is amazing