New fn for linking entities
No longer using execCommand because it was causing problems. Not only
the crazy cross-browser differences-- also because needing to select a
range in order to format it was a bit of a pain. That technique allowed
the undo stack to remain intact, but undoing the creation of a link
*only* removes the href from that link (at least in firefox)-- making it
not even really worth saving.

New command to link a range uses rangy's surroundContents function, and
also handles the cursor position itself instead of inserting/removing a
span.